Removing the is_stopped field
Moving the incrementing of the idle stop state to the end rather than
the beginning of the loop allows me to remove the is_stopped filed and
replace it with a calculated value.
is_stopped was storing the "were we stopped the last time
the update function ran".

By delaying all the state changes to self, we can outright avoid the
issue of EncoderState having some new values and some old values at the
same time.